+++ demo/views/notifications/toast-demo/toast-demo.js
@@ -29,11 +29,17 @@ class ToastDemo extends React.Component {
    * @method demo
    */
   get demo() {
+    let props = {}
+
+    if (this.value('dismissable')) {
+      props.onDismiss = this.action.bind(this, 'open', { target: { value: false } });
+    }
+
     return (
       <Toast
         as={ this.value('as') }
         open={ this.value('open') }
-        onDismiss={ this.action.bind(this, 'open', { target: { value: false } }) }
+        { ...props }
       >
         { this.value('text') || " " }
       </Toast>
@@ -75,11 +81,15 @@ class ToastDemo extends React.Component {
             value={ this.value('as') }
             onChange={ this.action.bind(this, 'as') }
           />
-
           <Checkbox
             label="Open"
             value={ this.value('open') }
             onChange={ this.action.bind(this, 'open') }
+          />
+          <Checkbox
+            label="Dismissable"
+            value={ this.value('dismissable') }
+            onChange={ this.action.bind(this, 'dismissable') }
           />
         </Row>
       </div>