+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * Allow Tables to shrink in size using the `shrink` prop.
 * Link component can now display with an icon.
 * Child components of Row can now use a `columnAlign` prop.
+* Toast onDismiss is now optional
 
 ## Bug Fixes